+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 
 * I bumped the major version for the new year.
 
+* I changed the build target from .net standard 2.0 to .net core 3.1
+
+* I added extensions to deal with efcore related startup tasks, such as migrations and seeding.
+
 ## 2020.5
 
 * I copied the code from the old CG.Linq.EntityFramework.